Enable discovery client and add initial package structure. Rename package name.
