+++ user-service/src/main/java/com/healthy/gym/user/UserServiceApplication.java
@@ -1,9 +1,11 @@
-package com.healthy.gym.userservice;
+package com.healthy.gym.user;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 
 @SpringBootApplication
+@EnableDiscoveryClient
 public class UserServiceApplication {
 
     public static void main(String[] args) {
