+++ user-service/src/main/java/com/healthy/gym/user/data/entity/UserEntity.java
@@ -1,0 +1,4 @@
+package com.healthy.gym.user.data.entity;
+
+public class UserEntity {
+}
