+++ user-service/src/main/java/com/healthy/gym/user/data/repository/UserDAO.java
@@ -1,0 +1,4 @@
+package com.healthy.gym.user.data.repository;
+
+public class UserDAO {
+}
