+++ user-service/src/main/java/com/healthy/gym/user/security/WebSecurity.java
@@ -1,0 +1,4 @@
+package com.healthy.gym.user.security;
+
+public class WebSecurity {
+}
