+++ user-service/src/main/java/com/healthy/gym/user/service/UserServiceImpl.java
@@ -1,0 +1,4 @@
+package com.healthy.gym.user.service;
+
+public class UserServiceImpl implements UserService{
+}
