+++ user-service/src/main/java/com/healthy/gym/user/shared/UserDTO.java
@@ -1,0 +1,4 @@
+package com.healthy.gym.user.shared;
+
+public class UserDTO {
+}
