+++ user-service/src/main/java/com/healthy/gym/user/ui/controller/UserController.java
@@ -1,0 +1,15 @@
+package com.healthy.gym.user.ui.controller;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/users")
+public class UserController {
+
+    @GetMapping("/status")
+    public String status(){
+        return "OK";
+    }
+}
