+++ user-service/src/main/java/com/healthy/gym/user/ui/model/request/CreateUserRequest.java
@@ -1,0 +1,4 @@
+package com.healthy.gym.user.ui.model.request;
+
+public class CreateUserRequest {
+}
