+++ user-service/src/main/java/com/healthy/gym/user/ui/model/response/CreateUserResponse.java
@@ -1,0 +1,4 @@
+package com.healthy.gym.user.ui.model.response;
+
+public class CreateUserResponse {
+}
